refactor: give the cross-window move its own module
The move subsystem is what this whole extraction was for: the in-flight
registration, the pending-content window it opens into, the abandoned-window
close, the recovery ladder with its machine-readable failure marker, and the
late-adopt probe. `MoveWebViewTarget`, the probe's attempt and delay constants,
`findWebViewAdoptedAfterTimeout`, `raiseMoveTarget`, `moveWebView`,
`readoptAfterFailedMove` and `recoverAfterFailedMove` move verbatim; the only
edits are the two `export`s the router needs and the imports each file no longer
uses.

The commands stay where they were registered. `MoveCommandName`,
`MOVE_COMMAND_DOCS` with its `x-experimental` markers, and the two handlers that
check their untyped network arguments all remain in the router, which now calls
in here for what they mean. `testingWebViewServiceRouter` keeps its four members
and its shape, assembled from the modules they now live in, so the tests that
import it need no edit.

The router is 873 lines, down from 1832.
